Finalze on TeacherDashboardHeader after A/B test
Consolidate TeacherDashboardHeader and TeacherDashboardHeaderWithButtons.
Keep only variant that features buttons, and a 'switch section' button to switch sections
